requirements.txt
opencv-python
numpy
pyserial
torch
torchvision
ultralytics
scikit-image
filterpy
